code/requirements.txt
# Python dependencies for WIG80 QuestDB Client
# ============================================

# Core async HTTP client
aiohttp>=3.8.0

# Date/time handling
python-dateutil>=2.8.0

# JSON processing (usually built-in)
# json5>=0.9.0  # Optional: for more flexible JSON parsing

# Data analysis and manipulation
pandas>=1.5.0
numpy>=1.21.0

# Data visualization (optional, for analysis scripts)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database connectivity (for PostgreSQL wire protocol)
asyncpg>=0.26.0
psycopg2-binary>=2.9.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Logging and monitoring
structlog>=22.0.0

# Testing (development dependencies)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.8.0

# Code quality (development dependencies)
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Jupyter notebooks support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# API development (optional)
fastapi>=0.85.0
uvicorn>=0.18.0

# Financial data processing (optional)
yfinance>=0.1.87
ta-lib>=0.4.0
